refactor(cli): carve plugin group into cli/groups/plugin.py
Move the plugin group, all seven subcommands (init/install/show/append/
scan/prompt/guard), and the plugin-library helpers out of the monolith.
Add cli/_shared.py for genuinely cross-module helpers and relocate
_contract_guarantee there (used by plugin + several commands). host
install imports _bootstrap_default_buckets and _install_one from the
plugin module. No behavior change (7 subcommands intact; tests pass).
